[1230] Rename '/new' route to '/search'
