+++ app/views/publish/providers/accredited_provider_search/new.html.erb
@@ -6,7 +6,7 @@
 
 <%= form_with(
         model: @accredited_provider_search_form,
-        url: publish_provider_recruitment_cycle_accredited_providers_path,
+        url: search_publish_provider_recruitment_cycle_accredited_providers_path,
         method: :post
       ) do |f| %>
