+++ app/views/publish/providers/accredited_provider_search/results.html.erb
@@ -7,7 +7,7 @@
 
 <%= form_with(
     model: @accredited_provider_select_form,
-    url: publish_provider_recruitment_cycle_accredited_providers_path,
+    url: search_publish_provider_recruitment_cycle_accredited_providers_path,
     method: :put
   ) do |f| %>
