+++ app/views/publish/providers/accredited_providers/index.html.erb
@@ -7,8 +7,8 @@
       </h1>
 
       <%= govuk_button_link_to("Add accredited provider",
-     new_publish_provider_recruitment_cycle_accredited_providers_path(provider_code: @provider.provider_code,
-                                                                      recruitment_cycle_year: @provider.recruitment_cycle_year)) %>
+     search_publish_provider_recruitment_cycle_accredited_providers_path(provider_code: @provider.provider_code,
+                                                                         recruitment_cycle_year: @provider.recruitment_cycle_year)) %>
     </div>
   </div>
