+++ config/routes/publish.rb
@@ -226,8 +226,15 @@ namespace :publish, as: :publish do
       end
 
       scope module: :providers do
-        resources :accredited_providers, only: [:index], path: '/accredited-providers'
-        resource :accredited_providers, only: %i[new create update], path: '/accredited-providers'
+        resources :accredited_providers, path: 'accredited-providers' do
+          get '/search', on: :collection, to: 'accredited_provider_search#new'
+          post '/search', on: :collection, to: 'accredited_provider_search#create'
+          put '/search', on: :collection, to: 'accredited_provider_search#update'
+
+          get '/check', on: :collection, to: 'accredited_providers/checks#show'
+          put '/check', on: :collection, to: 'accredited_providers/checks#update'
+        end
+
         resource :check_school, only: %i[show update], controller: 'schools_check', path: 'schools/check'
         resources :schools do
           member do
